+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - improve sub-command completion
     - complete sub-command from nested named imported modules
     - put sub-command description (`user-defined` or `module`)
+    - complete CLI option
 
 #### LSP